+++ HISTORY.md
@@ -3,9 +3,11 @@
 ## in development:
 * **Notable Changes:**
     * deprecated "expand (input) to (n) slots" command, because "set slot ... to ..." can now do this by passing a list as value
+    * merged both custom block slot event hat blocks into one "when slot ... gets ..."
 
 ### 2024-11-10
 * threads, objects, blocks: removed support for "expand" block (b/c reorganizing the whole feature)
+* threads, objects, blocks, byob: merged both ustom block slot event hat blocks into one "when slot ... gets ..."
 
 ### 2024-11-09
 * threads: enhanced "set slot ... to ..." command to also handle variadic inputs, setting them to a list of values